requirements.txt
aiogram>=3.3.0
telethon>=1.34.0
aiosqlite>=0.19.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pytz>=2024.1
peewee>=3.17.0
